pyproject.toml
[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "python-jose_aws-kms-extension"
version = "1.0.0"
description = "AWS KMS extension for 'python-jose' library."

license = "Apache-2.0"

authors = [
    "Lynx Tech <lynx-tech@amazon.com>",
]

readme = ["README.md", "LICENSE"]

packages = [{include = "jose_aws_kms_extension"}]

[tool.poetry.dependencies]
# Compatible Python versions
python = ">=3.8,<4"
# Python JOSE
python-jose = {extras = ["cryptography"], version = "3.3.0"}
# Boto AWS SDK
boto3 = {extras = ["kms"], version = "^1.26.155"}
boto3-stubs = {extras = ["kms"], version = "^1.26.155"}

[tool.poetry.group.dev.dependencies]
# PyTest (Test runner)
pytest = "^7.1.2"
pytest-cov = "^3.0"
# MyPy (for type checking)
mypy = "^1.3.0"
# Flake8 (For style checks)
pyproject-flake8 = "^5.0.4"
# Poe the Poet (For enhancing Peotry capabilities)
poethepoet = "^0.20.0"
# IPython (Interactive Python shell)
ipython = "^8.12"

[tool.mypy]
# See https://mypy.readthedocs.io/en/latest/config_file.html for more mypy options.

# Enables the type-checker on the interior of functions without type annotations.
check_untyped_defs = true

# Displaying specific error codes makes it easier to silence specific errors
# See also https://mypy.readthedocs.io/en/latest/error_codes.html
show_error_codes = true

# Show source code snippets and location markers in error messages
pretty = true

# Suppresses errors about packages which do not implement type-hint sharing.
# See also https://mypy.readthedocs.io/en/latest/running_mypy.html#missing-imports.
ignore_missing_imports = true

# Disallows defining functions without type annotations or with incomplete type annotations.
# See also https://mypy.readthedocs.io/en/stable/config_file.html#confval-disallow_untyped_defs
disallow_untyped_defs = true

[tool.flake8]
max-line-length = 120

[tool.poe.tasks]
lock = "poetry lock"
install = "poetry install"
build = "poetry build"
type-check = "poetry run mypy ."
style-check = "poetry run pflake8"
test = "poetry run pytest"
release = ["lock", "install", "type-check", "style-check", "test", "build"]
